+++ packages/@webex/internal-plugin-metrics/src/call-diagnostic/call-diagnostic-metrics-latencies.ts
@@ -308,15 +308,12 @@ export default class CallDiagnosticLatencies extends WebexPlugin {
     );
 
     // get the first timestamp
-    const mediaFlowStartedTimestamp = Math.min(
-      this.latencyTimestamps.get('client.media.rx.start'),
-      this.latencyTimestamps.get('client.media.tx.start')
-    );
+    const connectedMedia = this.latencyTimestamps.get('client.ice.end');
 
     const lobbyTime = this.getStayLobbyTime() || 0;
 
-    if (interstitialJoinClickTimestamp && mediaFlowStartedTimestamp) {
-      return mediaFlowStartedTimestamp - interstitialJoinClickTimestamp - lobbyTime;
+    if (interstitialJoinClickTimestamp && connectedMedia) {
+      return connectedMedia - interstitialJoinClickTimestamp - lobbyTime;
     }
 
     return undefined;